+++ src/renderer/components/LeftMenu.vue
@@ -4,7 +4,7 @@
     <!-- <i data-target="#cloudSettings" data-toggle="modal" class="fa fa-cloud-download fa-lg" title="Cloud Settings" /> -->
     <i v-if="!newVersion" v-on:click="openGithub" class="fa fa-github fa-2x" title="Github Repo"></i>
     <span v-if="newVersion" title="Newer version available" class="fa-stack">
-      <i class="fa fa-star fa-stack" style="color: gold; position: absolute; margin-top: 5px; margin-left: 20px;"></i>
+      <i class="fa fa-star fa-stack" style="font-size: 1em !important; color: gold; position: absolute; margin-top: 3px; margin-left: 18px;"></i>
       <i class="fa fa-github fa-stack-2x"></i>
     </span>
   </div>